agentic_platform/services/recipes/requirements.txt
# Core dependencies only
beautifulsoup4>=4.12.0
requests>=2.31.0
smolagents>=1.7.0
pika>=1.3.0
aio-pika>=9.0.0
jsonschema>=4.19.0

pytest>=7.4.3
pytest-mock>=3.14.0

pydantic>=2.5.3
datamodel-code-generator>=0.27.2

w3lib>=1.24.0
transformers>=4.48.3
recipe-scrapers>=15.4.0

google-generativeai>=0.8.4

python-dotenv>=1.0.1
aiohttp>=3.11.11
asyncio>=3.4.3


# Optional dependencies
# pydantic-mypy>=2.1.0
# pydantic-core>=2.10.0
# pydantic-core-extensions>=0.1.0
